requirements/base.txt
Django==1.10.2
flake8==3.0.4
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
channels==1.0.3
constantly==15.1.0
daphne==1.0.2
Django==1.10.2
django-filter==1.0.1
djangorestframework==3.5.3
flake8==3.0.4
incremental==16.10.1
Markdown==2.6.8
mccabe==0.5.3
msgpack-python==0.4.8
pbr==0.10.7
pep8-naming==0.4.1
psycopg2==2.6.2
pycodestyle==2.0.0
pyflakes==1.2.3
redis==2.10.5
six==1.10.0
Twisted==16.6.0
txaio==2.6.0
zope.interface==4.3.3
pillow==4.0.0
django-cors-headers==2.0.2
djoser==0.5.4
